Initial Trash and Pollution but mostly other stuff
Began by setting up items which trash will inherit from.  So far have done a fair bit of that.

Items have 3 parts, ItemData, ItemObject, and InventoryItem.

ItemData is a scriptable object with all the common data for every instance of the item type.

ItemObject is the 3D object of a specific instance of an item.  It handles the movement, interaction, and other traits of the object in the environment.  Currently this mostly involves floating or sinking.

InventoryItem is the object representing an item in an inventory.  It can theoretically be rotated or flipped.

Included is the InventorySpace class which tracks spaces being occupied or not in a grid of a declared size, and can do some various functions comparing two of them.

There is also some initial work setting up an inventory grid before I stopped as that is a distinct task and not needed for the trash and pollution systems.

I also added an ocean manager class.  Currently, it's main function is setting the surface level of the ocean which is used to determine how high buoyant items float.  It also has a placeholder system create the surface of the water and have it move with waves.  Buoyant items do take into account the waves when determining what surface level is.  The surface also fallows the player as they move.