Initial Pollution cleanup
Pollution effects that can be cleaned up with the cleaner spray have been added.  Visual Effects of cleaner spray have been improved.

Currently there seems to be a bug that is likely with unity, where sometimes having either of the pollutant objects selected when an object in the scene has active status toggled will throw an error message about onGUI.  This occurs even when the Pollutant script is not added to them.