Swim and Camera Speed Fix
A few minor related fixes to make movement and camera speed more reasonable.

Firstly, AddForce should be used in FixedUpdate unless the force type is impulse.  If used outside FixedUpdate, it will just wait until the FixedUpdate to apply.

Secondly, AddForce will by default factor in the time to have the stated force be applied over the course of 1 second.  As such trying to manually multiply the vector by Time.deltaTime (or more appropriately Time.fixedDeltaTime) will just be squaring the time multiplier.

With those tweaks, the move speed could be adjusted to a more reasonable number and still feel fast enough to not be bothersome.

In addition, camera sensitivity variables were added for that will multiply the camera movement so that is can be adjusted to desired amounts.  At some point it might be worth having separate camera sensitivity options for gamepad and mouse.