Fixed up Cleaner Spray and set up Pollutant Spawning
Cleaner Spray has been cleaned up and the majority of the code split between the Sprayer and the CleanerSpray.  This should allow for having multiple types of sprayers and cleaners in the future.

Can theoretically swap between multiple cleaner types (each with their own ammount), though there is currently not any controls to do so, nor more than one type of cleaner.

Additionally, Polluter class has been created which can be added to an object to make it spawn pollution in specified random areas around it.

Player now has their mass adjusted based on the combination of a base weight, the items they are carrying, sprayer weight, and how much cleaner they have.  To avoid significant slowdown from items, a multiplier for the inventory weight has been added, currently using a tenth of the total inventory weight.

An Audio Mixer has been set up, along with the code to adjust volume or mute of various types (Master, Music, Ambient, SFX, & UI).  Items will play a SFX on pickup if one is assigned to them (in theory).