Re-fixed cleaning for the swim test scene.
Another commit seemed to have undone the fixes to make cleaning work in the SwimTest scene (mainly relating to properly setting up the sprayer, and replacing the pollutants with prefab versions), as well as adding colliders to the player.

Additionally, tweaked the swimming sprint to work.  This involved setting the sprint button to not overlap with the swim up button input, assigning the sprint function to the sprint button, and fixing the method to ensure the player is under the water (rather than over) and setting the speed boost to add force as an impulse rather than the default.